+++ Src/Library/Auth/UserPrivileges.cs
@@ -3,7 +3,7 @@
 namespace FastEndpoints;
 
 /// <summary>
-/// the priviledges of the user which will be embedded in the jwt
+/// the priviledges of the user which will be embedded in the jwt or cookie
 /// </summary>
 public sealed class UserPrivileges
 {
@@ -21,4 +21,12 @@ public sealed class UserPrivileges
     /// allowed permissions for the user
     /// </summary>
     public List<string> Permissions { get; } = new();
+
+    /// <summary>
+    /// shortcut for adding a new <see cref="Claim"/> to the claim list for the given claim type and value
+    /// </summary>
+    /// <param name="claimType">the claim type to add</param>
+    public string this[string claimType] {
+        set => Claims.Add(new Claim(claimType, value));
+    }
 }